test: add fuzzer test
