+++ pkg/controller/direct/tags/mapper.generated.go
@@ -18,59 +18,3 @@
 // proto.service: google.cloud.resourcemanager.v3
 
 package tags
-
-import (
-	pb "cloud.google.com/go/resourcemanager/apiv3/resourcemanagerpb"
-	krm "github.com/GoogleCloudPlatform/k8s-config-connector/apis/tags/v1alpha1"
-	v1alpha1 "github.com/GoogleCloudPlatform/k8s-config-connector/pkg/clients/generated/apis/k8s/v1alpha1"
-	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct"
-)
-
-func TagsLocationTagBindingObservedState_FromProto(mapCtx *direct.MapContext, in *pb.TagBinding) *krm.TagsLocationTagBindingObservedState {
-	if in == nil {
-		return nil
-	}
-	out := &krm.TagsLocationTagBindingObservedState{}
-	// MISSING: Name
-	// MISSING: TagValueNamespacedName
-	return out
-}
-func TagsLocationTagBindingObservedState_ToProto(mapCtx *direct.MapContext, in *krm.TagsLocationTagBindingObservedState) *pb.TagBinding {
-	if in == nil {
-		return nil
-	}
-	out := &pb.TagBinding{}
-	// MISSING: Name
-	// MISSING: TagValueNamespacedName
-	return out
-}
-func TagsLocationTagBindingSpec_FromProto(mapCtx *direct.MapContext, in *pb.TagBinding) *krm.TagsLocationTagBindingSpec {
-	if in == nil {
-		return nil
-	}
-	out := &krm.TagsLocationTagBindingSpec{}
-	// MISSING: Name
-	if in.GetParent() != "" {
-		out.ParentRef = &v1alpha1.ResourceRef{External: in.GetParent()}
-	}
-	if in.GetTagValue() != "" {
-		out.TagValueRef = &krm.TagValueRef{External: in.GetTagValue()}
-	}
-	// MISSING: TagValueNamespacedName
-	return out
-}
-func TagsLocationTagBindingSpec_ToProto(mapCtx *direct.MapContext, in *krm.TagsLocationTagBindingSpec) *pb.TagBinding {
-	if in == nil {
-		return nil
-	}
-	out := &pb.TagBinding{}
-	// MISSING: Name
-	if in.ParentRef != nil {
-		out.Parent = in.ParentRef.External
-	}
-	if in.TagValueRef != nil {
-		out.TagValue = in.TagValueRef.External
-	}
-	// MISSING: TagValueNamespacedName
-	return out
-}
